requirements.txt
aiofiles==0.6.0
arrow==1.1.0
binaryornot==0.4.4
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
cookiecutter==1.7.3
dnspython==2.1.0
fastapi==0.63.0
h11==0.12.0
httpcore==0.12.3
httpx==0.17.1
idna==2.10
Jinja2==3.0.1
jinja2-time==0.2.0
loguru==0.5.3
MarkupSafe==2.0.1
motor==2.4.0
nb-cli==0.4.2
nonebot-adapter-cqhttp==2.0.0a13
nonebot-plugin-test==0.2.0
nonebot2==2.0.0a13.post1
poyo==0.5.0
prompt-toolkit==1.0.14
pydantic==1.8.2
pyfiglet==0.8.post1
Pygments==2.9.0
pygtrie==2.4.2
PyInquirer==1.0.3
pymongo==3.11.4
python-dateutil==2.8.1
python-dotenv==0.17.1
python-engineio==3.14.2
python-slugify==5.0.2
python-socketio==4.6.1
PyYAML==5.4.1
regex==2021.4.4
requests==2.25.1
rfc3986==1.5.0
six==1.16.0
sniffio==1.2.0
starlette==0.13.6
text-unidecode==1.3
tomlkit==0.7.2
typing-extensions==3.10.0.0
urllib3==1.26.4
uvicorn==0.13.4
watchgod==0.7
wcwidth==0.2.5
websockets==8.1
win32-setctime==1.0.3
wincertstore==0.2
